+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Category;
 use App\Models\SubCategory;
+use App\Models\ChildCategory;
 use Illuminate\Support\Str;
 
 class SubCategoryController extends Controller
@@ -99,6 +100,10 @@ class SubCategoryController extends Controller
     public function destroy(string $id)
     {
         $subCategory = SubCategory::findOrFail($id);
+        $childCategory = ChildCategory::where('sub_category_id', $subCategory->id)->count();
+        if($childCategory > 0 ) {
+            return response(['status' => 'error', 'message' => 'The item has a childCategory,first deleted childCAtegory related this item!']);
+        }
         $subCategory->delete();
 
         return response(['status' => 'success', 'message' => 'SubCategory deleted successfully']);
